+++ apps/3001-shop/cypress/e2e/app.cy.ts
@@ -6,7 +6,7 @@ describe('3001-shop/', () => {
   });
 
   describe('Warmup Next', () => {
-    it('warms pages concurrently', () => {
+    xit('warms pages concurrently', () => {
       const urls = [
         '/shop',
         '/checkout',