+++ packages/enhanced/src/schemas/container/ModuleFederationPlugin.ts
@@ -669,10 +669,6 @@ export default {
           description: 'Disable hot types reload.',
           type: 'boolean',
         },
-        injectWebClient: {
-          description: 'whether inject websocket client.',
-          type: 'boolean',
-        },
       },
     },
     AsyncBoundaryOptions: {