+++ packages/sdk/src/types/plugins/ModuleFederationPlugin.ts
@@ -128,7 +128,7 @@ export interface PluginManifestOptions {
 export interface PluginDevOptions {
   disableLiveReload?: boolean;
   disableHotTypesReload?: boolean;
-  injectWebClient?: boolean;
+  disableDynamicRemoteTypeHints?: boolean;
 }
 
 export interface DtsHostOptions {