+++ packages/webpack-bundler-runtime/src/constant.ts
@@ -1,5 +1,3 @@
-export const ENCODE_NAME_PREFIX = 'ENCODE_NAME_PREFIX';
-
 // TODO: support rspack
 export const FEDERATION_SUPPORTED_TYPES_CJS = 'cjs:webpack';
 export const FEDERATION_SUPPORTED_TYPES_CJS_MAP = {
@@ -10,3 +8,5 @@ export const FEDERATION_SUPPORTED_TYPES = [
   'script',
   FEDERATION_SUPPORTED_TYPES_CJS,
 ];
+
+export { ENCODE_NAME_PREFIX } from '@module-federation/sdk';