+++ packages/webpack-bundler-runtime/src/remotes.ts
@@ -1,8 +1,8 @@
 import { attachShareScopeMap } from './attachShareScopeMap';
 import type { RemoteEntryExports } from './types';
 import { RemotesOptions } from './types';
-import { decodeName } from '@module-federation/sdk';
 import { ENCODE_NAME_PREFIX, FEDERATION_SUPPORTED_TYPES } from './constant';
+import { decodeName } from '@module-federation/sdk';
 
 export function remotes(options: RemotesOptions) {
   const {