+++ packages/manifest/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @module-federation/manifest
 
+## 0.1.21
+
+### Patch Changes
+
+- Updated dependencies [a50c459]
+- Updated dependencies [88900ad]
+- Updated dependencies [4cddb8a]
+- Updated dependencies [6ee10aa]
+- Updated dependencies [1e93c5e]
+  - @module-federation/dts-plugin@0.1.21
+  - @module-federation/sdk@0.1.21
+  - @module-federation/managers@0.1.21
+
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies [685c607]
+- Updated dependencies [34f9498]
+- Updated dependencies [e8e0969]
+- Updated dependencies [349c381]
+  - @module-federation/dts-plugin@0.1.20
+  - @module-federation/sdk@0.1.20
+  - @module-federation/managers@0.1.20
+
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [031454d]
+- Updated dependencies [b0a31a7]
+- Updated dependencies [a2bfb9b]
+  - @module-federation/sdk@0.1.19
+  - @module-federation/dts-plugin@0.1.19
+  - @module-federation/managers@0.1.19
+
 ## 0.1.18
 
 ### Patch Changes