+++ packages/native-federation-tests/CHANGELOG.md
@@ -1,5 +1,11 @@
 # [1.0.0-canary.1](https://github.com/module-federation/core/compare/native-federation-tests-0.2.1...native-federation-tests-1.0.0-canary.1) (2023-11-06)
 
+## 0.5.1
+
+### Patch Changes
+
+- 3082718: ability to use relative paths
+
 ## 0.5.0
 
 ### Minor Changes