+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,49 @@
 # [8.1.0-canary.7](https://github.com/module-federation/core/compare/nextjs-mf-8.1.0-canary.6...nextjs-mf-8.1.0-canary.7) (2023-11-21)
 
+## 8.3.22
+
+### Patch Changes
+
+- Updated dependencies [88900ad]
+- Updated dependencies [1e93c5e]
+  - @module-federation/sdk@0.1.21
+  - @module-federation/enhanced@0.1.21
+  - @module-federation/node@2.2.12
+  - @module-federation/runtime@0.1.21
+  - @module-federation/utilities@3.0.26
+
+## 8.3.21
+
+### Patch Changes
+
+- 0f4d9e2: Do not crash on share resolver if host is not next
+- Updated dependencies [766de80]
+- Updated dependencies [652c8a2]
+- Updated dependencies [685c607]
+- Updated dependencies [d5ac542]
+- Updated dependencies [05c43f3]
+- Updated dependencies [e8e0969]
+- Updated dependencies [349c381]
+  - @module-federation/enhanced@0.1.20
+  - @module-federation/runtime@0.1.20
+  - @module-federation/sdk@0.1.20
+  - @module-federation/node@2.2.11
+  - @module-federation/utilities@3.0.25
+
+## 8.3.20
+
+### Patch Changes
+
+- Updated dependencies [031454d]
+- Updated dependencies [e1518e2]
+- Updated dependencies [b0a31a7]
+- Updated dependencies [a2bfb9b]
+  - @module-federation/runtime@0.1.19
+  - @module-federation/sdk@0.1.19
+  - @module-federation/enhanced@0.1.19
+  - @module-federation/node@2.2.10
+  - @module-federation/utilities@3.0.24
+
 ## 8.3.19
 
 ### Patch Changes